Added a new range-based signature option for event callbacks.
Instead of passing a function `func(E)` that is called for every event on the stream, EventRange<E>  allows to iterate them manually.
This allows client code to add node-level parallelization with parallel_for or parallel_reduce.

It's supported by existing Observe and MakeContinuation, and also by a new operation `Process`, which is a more generic form of Transform/Filter.
The signature of process callbacks is `void(EventRange<int> range, EventInserter<float> out)`. out is a back_insert_iterator.